Apply spread management to lptoken price

+++ contracts/interfaces/IPerpetualMarketCore.sol
@@ -11,11 +11,11 @@ interface IPerpetualMarketCore {
         int128[2] amountsFundingPaidPerPosition;
     }
 
-    function initialize(uint128 _depositAmount, int128 _initialFundingRate) external returns (uint128 mintAmount);
+    function initialize(uint256 _depositAmount, int256 _initialFundingRate) external returns (uint256 mintAmount);
 
-    function deposit(uint128 _depositAmount) external returns (uint256 mintAmount);
+    function deposit(uint256 _depositAmount) external returns (uint256 mintAmount);
 
-    function withdraw(uint128 _withdrawnAmount) external returns (uint256 burnAmount);
+    function withdraw(uint256 _withdrawnAmount) external returns (uint256 burnAmount);
 
     function addLiquidity(uint256 _amount) external;
